Refactor MDNS functions to separate file.
